requirements.txt
torch
langchain
langchain-community
pandas
chromadb
langchain-chroma
langchain-openai
langchain-text-splitters
python-dotenv
langchain-experimental
flask
docling
markdown